requirements.txt
python-dotenv==1.0.1
requests==2.32.0
chromadb==0.5.5
pydantic==2.8.2
sentence-transformers==3.0.1
ibm_watsonx_ai==1.1.7